+++ src/routers/admin.js
@@ -28,6 +28,7 @@ module.exports = function(app) {
 
     router.get('/system-information', Controller.systemInformation);
 
+    router.get('/show-config', Controller.showConfig);
 
     app.use('/admin', router);